Изменяет модульность проекта и убирает баги с загрузкой публикации
